Move max tick resolution calculation to state
